Add license confirmation before project download
